Rename per-layer embedding module
